Integrated akulen's changes
