+++ awesome/src/awesome.rs
@@ -74,7 +74,7 @@ pub fn init(lua: &Lua) -> rlua::Result<()> {
     method_setup(lua, &awesome_table)?;
     property_setup(lua, &awesome_table)?;
     let globals = lua.globals();
-    
+
     // Fixups: Add string.wlen
     let global_string: Table = globals.get("string")?;
     global_string.set("wlen", lua.create_function(wlen)?)?;
@@ -112,7 +112,7 @@ fn method_setup<'lua>(lua: &'lua Lua, awesome_table: &Table<'lua>) -> rlua::Resu
     awesome_table.set("pixbuf_to_surface", lua.create_function(pixbuf_to_surface)?)?;
     awesome_table.set("sync", lua.create_function(sync)?)?;
     awesome_table.set("exec", lua.create_function(exec)?)?;
-    awesome_table.set("spawn", lua.create_function(super::dummy)?)?;
+    awesome_table.set("spawn", lua.create_function(::objects::dummy)?)?;
     awesome_table.set("kill", lua.create_function(kill)?)?;
     awesome_table.set("quit", lua.create_function(quit)?)
 }
